fix(bench): torajs-str bench matches StrBlock u32 length signature
StrBlock::alloc / as_bytes_mut take u32 since the length-field port;
the criterion bench still passed u64 and cargo build --benches failed.
Not caught by pre-flight (--examples --tests does not build benches).

+++ crates/torajs-str/benches/str.rs
@@ -9,8 +9,8 @@ use torajs_bench::{Bench, bench_group, bench_main};
 use torajs_str::{__torajs_str_eq, __torajs_str_free, __torajs_str_slice, StrBlock};
 
 fn make_str(payload: &[u8]) -> *mut u8 {
-    let mut b = StrBlock::alloc(payload.len() as u64);
-    let dst = unsafe { b.as_bytes_mut(payload.len() as u64) };
+    let mut b = StrBlock::alloc(payload.len() as u32);
+    let dst = unsafe { b.as_bytes_mut(payload.len() as u32) };
     dst.copy_from_slice(payload);
     b.into_raw()
 }
